Adds function to apply Eq 46 from Gumerov paper
That's doi: 10.1137/S1064827501399705.

Rather than go for a sparse matrix, I've directly created a function to
apply the matrix to a two-column matrix containing the vector potentials
(Φ, Ψ). This is easier since the two columns of the matrix will not be
contiguous in practice. So viewing the operation as a sparse matrix
times a dense contiguous vector is not possible.